requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
PyJWT>=2.8.0
python-multipart>=0.0.9
